spilt maps components
